fix: make Not Found Page globally, move Home Page to `main` route group
