+++ app/(main)/page.tsx
@@ -1,0 +1,5 @@
+import MapDashboard from '@/components/map-dashboard'
+
+export default function Home() {
+  return <MapDashboard />
+}
