--- app/page.tsx
@@ -1,16 +1,0 @@
-import { Navbar } from '@/components/navbar'
-import MapDashboard from '../components/map-dashboard'
-
-export default function Home() {
-  return (
-    <>
-      <header>
-        <Navbar />
-      </header>
-
-      <main>
-        <MapDashboard />
-      </main>
-    </>
-  )
-}
